django-api/requirements.txt
asgiref==3.7.2
colorama==0.4.6
Django==4.0
django-filter==23.5
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
drf-nested-routers==0.93.5
exceptiongroup==1.2.1
iniconfig==2.0.0
packaging==24.0
pillow==10.2.0
pluggy==1.5.0
psycopg2-binary==2.9.9
PyJWT==2.8.0
pytest==8.2.0
pytest-django==4.8.0
python-dotenv==1.0.1
pytz==2024.1
sqlparse==0.4.4
tomli==2.0.1
typing_extensions==4.12.2
tzdata==2024.1